Add quiz generator to utility file
